requirements.txt
streamlit
streamlit-chat
python-dotenv
langchain
openai
huggingface_hub